pyproject.toml
[project]
name = "twitter-automation-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dotenv>=0.9.9",
    "feedparser>=6.0.12",
    "langchain>=1.2.13",
    "langchain-ollama>=1.0.1",
    "langchain-openai>=1.1.11",
    "lxml-html-clean>=0.4.4",
    "newspaper3k>=0.2.8",
    "psycopg2-binary>=2.9.11",
    "sentence-transformers>=5.3.0",
    "tweepy>=4.16.0",
]
